Update 119. Pascal's Triangle II.java

+++ 119. Pascal's Triangle II.java
@@ -1,9 +1,9 @@
 class Solution {
     public List<Integer> getRow(int rowIndex) {
         List<Integer> row = new ArrayList<>();
-        int value=1;
+        long value=1;
         for (int i=0;i<=rowIndex;i++) {
-            row.add(value);
+            row.add((int)value);
             value=value*(rowIndex-i)/(i+1);
         }
         return row;
